SlackLine/requirements.txt
slack-sdk>=3.27.0
line-bot-sdk>=3.5.0

